webapp/backend/requirements.txt
fastapi
uvicorn
motor
pydantic[email]
pydantic-settings
passlib[bcrypt]
pyjwt
python-dotenv
